Отдельный презентер для точек маршрута
